tests/selenium/requirements.txt
# Selenium UI Testing Dependencies
selenium==4.18.1
pytest==8.0.0
pytest-html==4.1.1
pytest-xdist==3.5.0
python-dotenv==1.0.1
Pillow==10.2.0
requests==2.31.0
webdriver-manager==4.0.1
allure-pytest==2.13.2
